mypy: function return types marked as optional
Only methods can have optional return types, but it has to be applied to function
as well as method, otherwise mypy complains.

I considered making a special return type for constructors/destructors, but it
seemed to just make things too messy. Time will tell if this was the right
decision.